Changed amount to decimal

+++ BedeSlots/BedeSlots.Services.Data/Contracts/ITransactionService.cs
@@ -1,5 +1,4 @@
 ﻿using BedeSlots.Data.Models;
-using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
@@ -13,6 +12,6 @@ namespace BedeSlots.Services.Data.Contracts
 
         Task<Transaction> GetTransactionAsync(int id);
 
-        Transaction CreateTransaction(TransactionType type, string userId, int cardId,double depositAmount);
+        Transaction CreateTransaction(TransactionType type, string userId, int cardId, decimal depositAmount);
     }
 }
